refactor types for 32bit support

+++ client/client_ad_integration_test.go
@@ -45,7 +45,7 @@ func TestClient_GetServiceTicket_AD(t *testing.T) {
 		t.Fatalf("Error getting service ticket: %v\n", err)
 	}
 	assert.Equal(t, spn, tkt.SName.GetPrincipalNameString())
-	assert.Equal(t, 18, key.KeyType)
+	assert.Equal(t, int32(18), key.KeyType)
 }
 
 func TestClient_SuccessfulLogin_AD_TRUST_USER_DOMAIN(t *testing.T) {
